+++ news_pages/news_109225887.php
@@ -15,7 +15,7 @@ The Valve Cyber Café License Server is now available to subscribers of the Valv
 <br/><br/></p>
 <p>
 
-A <a href="/index.php?area=productnews_CSS">Counter-Strike: Source beta news page</a> has been added to the site, to provide you with the latest news on the beta.  Also be sure to check out the <a href="/index.php?area=css_b1">Counter-Strike Source beta FAQ</a>, for the most commonly asked questions.<br/>
+A <a href="/index.php?area=productnews">Counter-Strike: Source beta news page</a> has been added to the site, to provide you with the latest news on the beta.  Also be sure to check out the <a href="/index.php?area=css_b1">Counter-Strike Source beta FAQ</a>, for the most commonly asked questions.<br/>
 <br/><br/></p>
 <p>