+++ news_pages/news_289.php
@@ -1,3 +1,3 @@
 <p>
 
-A <a href="/index.php?area=productnews_CSS">Counter-Strike: Source beta news page</a> has been added to the site, to provide you with the latest news on the beta.  Also be sure to check out the <a href="/index.php?area=css_b1">Counter-Strike Source beta FAQ</a>, for the most commonly asked questions.<br/></p>
+A <a href="/index.php?area=productnews">Counter-Strike: Source beta news page</a> has been added to the site, to provide you with the latest news on the beta.  Also be sure to check out the <a href="/index.php?area=css_b1">Counter-Strike Source beta FAQ</a>, for the most commonly asked questions.<br/></p>